+++ CHANGELOG.md
@@ -13,6 +13,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Changed
 
+- Added sourceType 'kotlin' to freemarker-doc-process-1-0.xsd
+
 ## [8.10.3] - 2024-10-26
 
 - [fj-doc-maven-plugin] goal init/add - support for base-json, base-yaml and base-kotlin example <https://github.com/fugerit-org/fj-doc/issues/231>